requirements.txt
scikit-build>=0.13.1
wheel>=0.37.1
numpy==1.19.4
#opencv-python>=4.5.0
#boto3>=1.23.10
PyYAML>=6.0
python-dotenv>=0.19.0
